+++ pumpkin-crates/core/src/checkers/support.rs
@@ -29,11 +29,14 @@ pub trait SupportGenerator: Clone + Debug {
     );
 
     /// Returns true if the support is a solution to the constraint.
+    ///
+    /// Called with the support generated by [`SupportGenerator::support`].
     fn is_solution(&self, support: &Support<Self::Value>) -> bool;
 }
 
 /// A value that may be used in a [`Support`].
 pub trait SupportValue: Clone + Debug {
+    /// Returns `true` if `self` is in the given domain.
     fn is_in(&self, domain: DomainId, domains: &Domains<'_>) -> bool;
 
     /// If the value is an integer, we can cache it to prevent recreating supports for the same
@@ -112,11 +115,20 @@ impl<Value: SupportValue> Support<Value> {
 #[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct UnsupportedValue(pub(crate) i32);
 
+/// Implementors know how to unpack an [`UnsupportedValue`].
 pub trait UnpackUnsupportedValue {
     /// Turn the unsupported value into an item in the domain of [`self`] (the variable).
     fn unpack(&self, unsupported_value: UnsupportedValue) -> i32;
 }
 
+impl UnpackUnsupportedValue for i32 {
+    fn unpack(&self, UnsupportedValue(value): UnsupportedValue) -> i32 {
+        assert_eq!(value, *self);
+        value
+    }
+}
+
+/// A trait to identify types that can serve as variables in a support.
 pub trait SupportsValue<Value = i32>: UnpackUnsupportedValue {
     /// Add the assignment `self = value` to the `support`.
     fn assign(&self, value: Value, support: &mut Support<Value>);
@@ -127,13 +139,6 @@ pub trait SupportsValue<Value = i32>: UnpackUnsupportedValue {
     ///
     /// Panics if the support has no value for this variable.
     fn support_value(&self, support: &Support<Value>) -> Value;
-}
-
-impl UnpackUnsupportedValue for i32 {
-    fn unpack(&self, UnsupportedValue(value): UnsupportedValue) -> i32 {
-        assert_eq!(value, *self);
-        value
-    }
 }
 
 impl SupportsValue<i32> for i32 {
